[PHP] CRUD Sản phẩm và biến thể sản phẩm và bắt lỗi xóa user, thể loại , đặt hàng  , sản phẩm (N3-62)

+++ Admin/src/category/CategoryDetele.php
@@ -2,10 +2,20 @@
  
 if (isset($_POST['DeleteCate'])) {
     $id = $_GET ['id'];  
-    $deletecate=new categories();
-    $DeleteCategory=$deletecate->delete($id);
-    header("location:/index.php?pages=admin&layout=home&modulde=category&action=list");
-    ob_end_flush();
+    if(!empty($id)){
+        try {
+            $deletecate=new categories();
+            $DeleteCategory=$deletecate->delete($id);
+            header("location:/index.php?pages=admin&layout=home&modulde=category&action=list");
+            ob_end_flush();
+            echo "Không bao giờ in ra dòng này vì có lỗi trước đó.";
+        } catch (Exception $e) {
+           // lỗi in ra dòng này 
+            $mess ="bị dính khóa ngoại vui lòng xóa ngoại trước" ;
+        }
+
+    }
+   
 }
 ?>
 <div class="content-body" style="min-height: 780px;">
@@ -20,6 +30,7 @@ if (isset($_POST['DeleteCate'])) {
                             <h3>Xóa Phân Loại</h3>
                             <form class="form-valide" action="#" method="post" novalidate="novalidate">
                                 <p>Bạn chắc chắn xóa ?</p>
+                                <p style="color: red;"> <?php echo $mess ?? "" ?></p>
                                 <div class="form-group row">
                                     <div class="col-lg-8 ml-auto">
                                         <button type="submit" class="btn btn-primary " name="DeleteCate">Xác Nhận</button>
